# Convert to int for faster mod