@mhostetter 2025-04-19
Thanks for running these tests. I'm happy to add that back in. I'll wait for the CI to pass and then merge.